Migrated the RegisterServiceImpl class and its methods to AuthServiceImpl.
