+++ nexus/src/main/java/com/example/nexus/controller/AuthController.java
@@ -3,7 +3,6 @@ package com.example.nexus.controller;
 import com.example.nexus.model.payload.request.AuthenticationRequest;
 import com.example.nexus.model.payload.request.RegisterRequest;
 import com.example.nexus.service.AuthService;
-import com.example.nexus.service.RegisterService;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
@@ -20,7 +19,6 @@ import org.springframework.web.bind.annotation.RestController;
 @Slf4j
 public class AuthController {
     private final AuthService authService;
-    private final RegisterService registerService;
 
     @PostMapping("/token")
     public ResponseEntity<?> login(@RequestBody AuthenticationRequest authenticationRequest) {
@@ -37,7 +35,7 @@ public class AuthController {
 
     @PostMapping("/registration")
     ResponseEntity<?> register(@RequestBody RegisterRequest registerRequest) {
-        var bodyContent = registerService.registerUser(registerRequest);
+        var bodyContent = authService.registerUser(registerRequest);
 
         return ResponseEntity.status(HttpStatus.CREATED).body(bodyContent);
     }
