+++ nexus/src/main/java/com/example/nexus/service/AuthService.java
@@ -1,7 +1,10 @@
 package com.example.nexus.service;
 
+import com.example.nexus.model.entity.Profile;
 import com.example.nexus.model.payload.request.AuthenticationRequest;
+import com.example.nexus.model.payload.request.RegisterRequest;
 
 public interface AuthService {
     String login(AuthenticationRequest request);
+    Profile registerUser(RegisterRequest registerRequest);
 }
