+++ nexus/src/main/java/com/example/nexus/service/AuthServiceImpl.java
@@ -1,9 +1,18 @@
 package com.example.nexus.service;
 
 import com.example.nexus.constant.MessageConstants;
+import com.example.nexus.constant.RegexConstants;
+import com.example.nexus.constant.RoleConstants;
+import com.example.nexus.exception.NotFoundException;
 import com.example.nexus.exception.UnauthorizedException;
+import com.example.nexus.exception.UserAlreadyExistsException;
+import com.example.nexus.mapper.RegisterMapper;
+import com.example.nexus.model.entity.Profile;
 import com.example.nexus.model.entity.User;
 import com.example.nexus.model.payload.request.AuthenticationRequest;
+import com.example.nexus.model.payload.request.RegisterRequest;
+import com.example.nexus.repository.ProfileRepository;
+import com.example.nexus.repository.RoleRepository;
 import com.example.nexus.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.authentication.AuthenticationManager;
@@ -18,6 +27,9 @@ public class AuthServiceImpl implements AuthService {
     private final AuthenticationManager authenticationManager;
     private final UserRepository userRepository;
     private final PasswordEncoder passwordEncoder;
+    private final ProfileRepository profileRepository;
+    private final RoleRepository roleRepository;
+    private final RegisterMapper registerMapper;
 
     @Override
     public String login(AuthenticationRequest request) {
@@ -37,10 +49,43 @@ public class AuthServiceImpl implements AuthService {
         return "token";
     }
 
+    @Override
+    public Profile registerUser(RegisterRequest registerRequest) {
+
+        validateRequestedPassword(registerRequest);
+
+        var newUser = registerMapper.mapUser(registerRequest);
+        isUserNameTaken(newUser.getUsername());
+        newUser.setPassword(passwordEncoder.encode(registerRequest.password()));
+        roleRepository.findByName(RoleConstants.USER).
+                ifPresentOrElse(userRole -> newUser.getRoles().add(userRole),
+                        () -> {throw new NotFoundException(MessageConstants.ROLE_NOT_FOUNT);});
+
+        var newProfile = registerMapper.mapProfile(registerRequest);
+        newProfile.setUser(newUser);
+
+        return profileRepository.save(newProfile);
+    }
+
     private void validateUserPassword(User user, AuthenticationRequest request) {
         final var encodedRealPassword = user.getPassword();
         if (!this.passwordEncoder.matches(request.password(), encodedRealPassword)) {
             throw new UnauthorizedException(MessageConstants.INVALID_USERNAME_PASSWORD);
         }
+    }
+
+    private void validateRequestedPassword(RegisterRequest registerRequest) {
+        if (!registerRequest.password().matches(RegexConstants.PASSWORD_REQUIREMENT_REGEX)) {
+            throw new UnauthorizedException(MessageConstants.WRONG_PASSWORD_FORMAT);
+        }
+        if (!registerRequest.password().equals(registerRequest.confirmPassword())) {
+            throw new UnauthorizedException(MessageConstants.CONFIRM_PASSWORD_NOT_MATCHING);
+        }
+    }
+
+    private void isUserNameTaken(String username) {
+        userRepository.findByUsername(username).ifPresent(
+                user -> {throw new UserAlreadyExistsException(MessageConstants.USER_ALREADY_EXISTS);}
+        );
     }
 }
