--- nexus/src/main/java/com/example/nexus/service/RegisterService.java
@@ -1,8 +1,0 @@
-package com.example.nexus.service;
-
-import com.example.nexus.model.entity.Profile;
-import com.example.nexus.model.payload.request.RegisterRequest;
-
-public interface RegisterService {
-    Profile registerUser(RegisterRequest registerRequest);
-}
